@@ -11,10 +11,10 @@ description: A new Flutter project.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.5.0+38
+version: 1.5.8+46
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -24,27 +24,27 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   intl: ^0.17.0
-  video_player: '2.0.2'
+  video_player: ^2.1.1
   connectivity: '^3.0.2'
   chewie: ^1.0.0
-  firebase_auth: ^1.0.1
+  firebase_auth: ^1.1.1
   firebase_database: ^6.1.0
-  firebase_analytics: ^7.1.1
-  firebase_crashlytics: ^1.0.0
+  firebase_analytics: ^8.0.3
+  firebase_crashlytics: ^2.0.0
   http: ^0.13.0
   url_launcher: ^6.0.2
   flutter_launcher_icons: ^0.9.0
   modal_progress_hud: ^0.1.3
-  flutter_widget_from_html: ^0.6.0-rc.2021031201
-  in_app_purchase: ^0.5.1
+  flutter_widget_from_html: ^0.6.1
+  in_app_purchase: ^0.5.2
   flutter_icons: ^1.1.0
   package_info: ^2.0.0
   provider: ^5.0.0
-  transparent_image: ^2.0.0-nullsafety.0
-  cached_network_image: ^3.0.0-nullsafety
+  transparent_image: ^2.0.0
+  cached_network_image: ^3.0.0
 
 dev_dependencies:
-  test: ^1.15.2
+  test: ^1.16.0
   flutter_test:
     sdk: flutter
 
