@@ -7,24 +7,26 @@ class HomeAnimator extends StatefulWidget {
 
 class _HomeAnimatorState extends State<HomeAnimator>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
+  AnimationController? _controller;
 
   @override
   void initState() {
     _controller = AnimationController(
         duration: Duration(milliseconds: 1500), vsync: this);
-    _controller.forward();
+      _controller!.forward();
+
+
     super.initState();
   }
 
   @override
   void dispose() {
-    _controller.dispose();
+    _controller!.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    return Home(controller: _controller);
+    return Home(controller: _controller!);
   }
 }
