@@ -4,8 +4,7 @@ import 'package:evgeshayoga/provider/user_provider_model.dart';
 import 'package:evgeshayoga/ui/video_content/yoga_online_screen.dart';
 import 'package:evgeshayoga/utils/ProgressHUD.dart';
 import 'package:evgeshayoga/utils/style.dart';
-import 'package:firebase_auth/firebase_auth.dart' as fbauth;
-import 'package:firebase_database/firebase_database.dart';
+import 'package:firebase_auth/firebase_auth.dart' as fb;
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart' show PlatformException;
 import 'package:http/http.dart' as http;
@@ -18,11 +17,10 @@ class Login extends StatefulWidget {
 }
 
 class _LoginState extends State<Login> {
-  DatabaseReference databaseReference;
-  final fbauth.FirebaseAuth _auth = fbauth.FirebaseAuth.instance;
+  final fb.FirebaseAuth _auth = fb.FirebaseAuth.instance;
   final _loginFormKey = GlobalKey<FormState>();
   User user = User("", "", "", "");
-  String _loginAlert = "";
+  String? _loginAlert;
   bool _isInAsyncCall = false;
   bool _isTablet = false;
 
@@ -77,9 +75,9 @@ class _LoginState extends State<Login> {
                                     labelText: 'Email',
                                   ),
                                   onSaved: (value) =>
-                                      user.userEmail = value.trim(),
+                                      user.userEmail = value!.trim(),
                                   validator: (value) {
-                                    if (value.isEmpty) {
+                                    if (value!.isEmpty) {
                                       return 'Введите email';
                                     }
                                     return null;
@@ -93,13 +91,13 @@ class _LoginState extends State<Login> {
                                     labelText: 'Пароль',
                                   ),
                                   validator: (value) {
-                                    if (value.isEmpty) {
+                                    if (value!.isEmpty) {
                                       return 'Введите пароль';
                                     }
                                     return null;
                                   },
                                   onSaved: (value) =>
-                                      user.password = value.trim(),
+                                      user.password = value!.trim(),
                                 ),
                               ],
                             ),
@@ -141,7 +139,7 @@ class _LoginState extends State<Login> {
                       Padding(padding: new EdgeInsets.all(10.5)),
                       Center(
                         child: Text(
-                          _loginAlert,
+                          _loginAlert!,
                           style: TextStyle(
                             fontFamily: "Nunito",
                             color: Colors.red,
@@ -195,9 +193,9 @@ class _LoginState extends State<Login> {
                         hintText: 'Введите свой email',
                         labelText: 'Email',
                       ),
-                      onSaved: (value) => user.userEmail = value.trim(),
+                      onSaved: (value) => user.userEmail = value!.trim(),
                       validator: (value) {
-                        if (value.isEmpty) {
+                        if (value!.isEmpty) {
                           return 'Введите email';
                         }
                         return null;
@@ -212,12 +210,12 @@ class _LoginState extends State<Login> {
                         labelText: 'Пароль',
                       ),
                       validator: (value) {
-                        if (value.isEmpty) {
+                        if (value!.isEmpty) {
                           return 'Введите пароль';
                         }
                         return null;
                       },
-                      onSaved: (value) => user.password = value.trim(),
+                      onSaved: (value) => user.password = value!.trim(),
                     ),
                   ],
                 ),
@@ -268,7 +266,7 @@ class _LoginState extends State<Login> {
             child: Padding(
               padding: const EdgeInsets.all(18.0),
               child: Text(
-                _loginAlert,
+                _loginAlert!,
                 style: TextStyle(color: Colors.red),
               ),
             ),
@@ -326,9 +324,9 @@ class _LoginState extends State<Login> {
   }
 
   Future userLogIn() async {
-    if (_loginFormKey.currentState.validate()) {
-      _loginFormKey.currentState.save();
-      _loginFormKey.currentState.reset();
+    if (_loginFormKey.currentState!.validate()) {
+      _loginFormKey.currentState!.save();
+      _loginFormKey.currentState!.reset();
       try {
         _showProgressIndicator();
         var response = await http.post(
@@ -342,13 +340,13 @@ class _LoginState extends State<Login> {
         if (error != null) {
           throw new Exception(error);
         }
-        fbauth.UserCredential result = await _auth.signInWithEmailAndPassword(
+        fb.UserCredential result = await _auth.signInWithEmailAndPassword(
           email: user.userEmail,
           password: user.password,
         );
-        fbauth.User newUser = result.user;
+        fb.User? newUser = result.user;
         await Provider.of<UserProviderModel>(context, listen: false)
-            .login(newUser.uid);
+            .login(newUser!.uid);
         var router = new MaterialPageRoute(builder: (BuildContext context) {
           return YogaOnlineScreen(userUid: newUser.uid,);
         });
